fix a couple warnings; introduce c++17 and remove boost dependency in favor of spdlog and std::filesystem!

+++ renderlib/threading.h
@@ -15,4 +15,4 @@
 ///
 ///
 void
-parallel_for(size_t nb_elements, std::function<void(int start, int end)> functor, bool use_threads = true);
+parallel_for(size_t nb_elements, std::function<void(size_t start, size_t end)> functor, bool use_threads = true);
